requirements.txt
django<3
djangorestframework
django-filter
ipython
flake8
isort
pygments